Setup redux with mock data and make it work
