+++ actions/index.js
@@ -1,0 +1,4 @@
+export const loadProfile = profile => ({
+  type: "LOAD_PROFILE",
+  profile
+})
