+++ reducers/ProfileReducer.js
@@ -1,0 +1,10 @@
+const test = { test: "test"}
+
+export const profileReducer = (state = test, action) => {
+  switch (action.type) {
+    case "LOAD_PROFILE":
+      return action.profile;
+    default:
+      return state;
+  }
+};
