+++ reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import { profileReducer } from './ProfileReducer';
+
+const rootReducer = combineReducers({
+  profile: profileReducer
+})
+
+export default rootReducer;
